Benchmark (#1) (#133)
* Measuring time to create server providing pod and mark ready, hit rates..

Iss llm-d-incubation#102: Modified scenario for maintenance planning as suggested in discussion.

Adding GPU usage to logs, fixing hit rate computation

Fixes llm-d-incubation#136 and llm-d-incubation#137

llm-d-incubation#135: Initial work toward interrupting the benchmark for debugging.

* Standardized the return for pod wait to ScenarioResult

Now the return from the wait includes ScenarioResult and an error attribute.

In case of success, the ScenarioResult contains the provider pods and their hit/cold details, otherwise it contains both *some* of the pods that may have hit ready and the pods that never hit unready.

In case of failures, the intermediate results are printed in addition to the logs being dumped.

---------

Signed-off-by: aavarghese <avarghese@us.ibm.com>
Signed-off-by: Gloire Rubambiza <gloire@ibm.com>
Co-authored-by: Gloire Rubambiza <gloire@ibm.com>